extract runStatements and hasExplicitTransaction into migration/runner.ts
