+++ src/lib/orm/migration/apply.ts
@@ -1,8 +1,8 @@
-import type { SQL as SqlType, TransactionSQL } from "bun";
 import { SQL } from "bun";
 import { loadConfig } from "./config.js";
 import { loadOrmFromSchema } from "./discover.js";
-import { listMigrations, relativeFromCwd, splitStatements } from "./files.js";
+import { listMigrations, relativeFromCwd } from "./files.js";
+import { runMigrationSql } from "./runner.js";
 import {
   ensureMigrationStateTable,
   getMigrationTableName,
@@ -10,62 +10,6 @@ import {
   markAppliedMigration,
 } from "./state-table.js";
 import type { ApplyMigrationsInput } from "./types.js";
-
-async function runStatements(
-  runner: SqlType | TransactionSQL,
-  sqlText: string,
-) {
-  const statements = splitStatements(sqlText);
-
-  for (const statement of statements) {
-    await runner`${runner.unsafe(statement)}`;
-  }
-}
-
-function hasExplicitTransaction(sqlText: string) {
-  const statements = splitStatements(sqlText);
-
-  for (const statement of statements) {
-    const normalized = statement.trim().toUpperCase();
-
-    if (normalized === "BEGIN") {
-      return true;
-    }
-
-    if (normalized.startsWith("BEGIN ")) {
-      return true;
-    }
-
-    if (normalized === "START TRANSACTION") {
-      return true;
-    }
-
-    if (normalized.startsWith("COMMIT")) {
-      return true;
-    }
-
-    if (normalized.startsWith("ROLLBACK")) {
-      return true;
-    }
-  }
-
-  return false;
-}
-
-async function runMigrationSql(
-  sql: SQL,
-  sqlText: string,
-  transactional: boolean,
-) {
-  if (transactional && !hasExplicitTransaction(sqlText)) {
-    await sql.begin(async (tx) => {
-      await runStatements(tx, sqlText);
-    });
-    return;
-  }
-
-  await runStatements(sql, sqlText);
-}
 
 export async function applyMigrations(input?: ApplyMigrationsInput) {
   const cwd = input?.cwd ?? process.cwd();
